requirements.txt
streamlit
pandas
numpy
scikit-learn
pytesseract
Pillow
google-generativeai
requests
python-dotenv
fpdf
joblib
python-dotenv
pdfplumber